support default relabel config in validation
Prometheus has a default relabel config that is different
from the natural zero values for the individual types.

The current validations create a situation where a valid configuration
that is marshaled and unmarshaled (e.g. when generating a
ServiceMonitor config) is suddenly no longer valid.

This changes the validation to accept both the zero values (as that
makes the most sense when writing the config by hand) and the
default config as it is used by Prometheus (see also
https://github.com/prometheus/prometheus/blob/v2.33.3/model/relabel/relabel.go#L122-L130).

(cherry picked from commit 5f69d47b248ce413cc720d22d366ea158dbe0de4)